requirements.txt
python-fasthtml>=0.6.4
python-dotenv>=1.0.1
langchain-anthropic>=0.2.1
langchain-core>=0.3.5